requirements.txt
perplexityai>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0


